Update README.md
If your job involves sitting for long hours, it is essential to learn how to sit properly at work in order to avoid injury.

Computer Posture

Using a laptop is by far the worst type of gadgets for your back. If you are making use of a laptop which is placed on the table, you probably have the tendency to slouch in the direction of the screen. This type of sitting position damages your back. You will likely be at risk of lower back pain and a stiff neck should you continue with this computer posture. Make sure that your laptop screen and chair height are at the same level.

Adjust your chair

The majority of us have Executive chairs with adjustable heights in the office. However,some of us simply do not take the time to adjust the height of the chair, which result in needless strain on our back. Simply adjust the height of the chair in an effort to use your keyboard with your wrist and forearms at level with the floor.

Rest your feet on the floor

Your feet must be flat on the floor. In case you are unable to reach the floor, get a footrest, which allows your feet to rest at a comfortable level. Never sit with your legs crossed, since this can lead to posture-related problems.

Prevent phone strain

In case you spend plenty of time on the telephone, rather make use of a headset.

Make objects accessible

Place objects that you use often, like your telephone or stapler, close to you. Avoid continuously stretching or turning to get to things.

Making use of the keyboard

Your keyboard should be placed in front of you for easy typing. Make sure there is a small gap at the front of the desk for your wrists to rest between rounds of typing. Your wrists must be straight when working with a keyboard. Elbows should be kept vertical under your shoulder and right by your side.

Working with spectacles

Working with bifocal spectacles is not ideal for computer work. It is important that you can see the screen clearly without the need to raise or lower your head. If you are unable to work while wearing bifocals, you might need another type of spectacles. Seek advice from your optician if uncertain.

Position your screen at eye level

Your screen must be straight before you. A great guide would be to set the monitor about an arm's length away, ensuring that the top of the monitor is about at eye level. To accomplish this,place your monitor on a stand. When the screen is too low or too high,you will need to bend your neck,which can result in injury.

Keep your mouse close

Place you’re the mouse as close to you as you can. Using a mouse mat as well as a wrist pad can help to keep your wrist straight and prevent uncomfortable bending. If you aren't working with your keyboard, set it aside when you use your mouse a lot.

Hi I am from Sitsmart and I am here to post some really cool stuff for you. If you have any ideas or any request please contact me @- sales@sitsmart.co.uk you can also visit on my websites http://sitsmart.co.uk